This adds a transaction ID to the request.
This is the second part of having a transaction ID that goes though our
whole stack. This will let us trace the cause of requests, and their
consequences, and will hopefully make it easy for us to do things like
ascribe actions that reactor carries out to users in atom.